// src/types/api/service/models/LoginDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * MediaBee Service API
 * MediaBee Service API
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface LoginDto
 */
export interface LoginDto {
    /**
     * 이메일(ID)
     * @type {string}
     * @memberof LoginDto
     */
    email: string;
    /**
     * 비밀번호
     * @type {string}
     * @memberof LoginDto
     */
    passwd: string;
}

/**
 * Check if a given object implements the LoginDto interface.
 */
export function instanceOfLoginDto(value: object): value is LoginDto {
    if (!('email' in value) || value['email'] === undefined) return false;
    if (!('passwd' in value) || value['passwd'] === undefined) return false;
    return true;
}

export function LoginDtoFromJSON(json: any): LoginDto {
    return LoginDtoFromJSONTyped(json, false);
}

export function LoginDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): LoginDto {
    if (json == null) {
        return json;
    }
    return {
        
        'email': json['email'],
        'passwd': json['passwd'],
    };
}

export function LoginDtoToJSON(value?: LoginDto | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'email': value['email'],
        'passwd': value['passwd'],
    };
}